requirements.txt
Django==4.2.4
django-restframework==0.0.1
djangorestframework==3.14.0
Pillow==10.0.0